requirements.txt
GitPython
python-dotenv
mistral
fastapi
pydantic
uvicorn